requirements.txt
aiohttp==3.8.1
psycopg2-binary==2.9.3
python-decouple==3.6
SQLAlchemy==1.4.41
beautifulsoup4==4.11.1
